Move `ReviewCommentsDB` to the `code_review.py`
